third_party/WebKit/LayoutTests/fast/dom/shadow/suppress-mutation-events-in-shadow-expected.txt

CONSOLE MESSAGE: DOMNodeInserted is deprecated. Use MutationObservers instead.
Test that mutation events are not triggered in shadow DOM

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mutationEventTriggered is false
PASS successfullyParsed is true

TEST COMPLETE
